Created a controller for issues/feedback
