+++ grails-app/controllers/cavamedia/VideoUploadController.groovy
@@ -2,11 +2,6 @@ package cavamedia
 
 import grails.converters.JSON
 import grails.util.Environment
-import grails.util.Holders
-import io.micronaut.http.client.HttpClient
-import io.micronaut.http.HttpRequest
-import io.micronaut.http.HttpResponse
-import io.micronaut.http.HttpStatus
 import io.swagger.annotations.Api
 import io.swagger.annotations.ApiImplicitParam
 import io.swagger.annotations.ApiImplicitParams
@@ -14,141 +9,19 @@ import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import io.swagger.annotations.ApiResponse
 import io.swagger.annotations.ApiResponses
-import org.springframework.beans.factory.annotation.Value
-
-// import io.swagger.v3.oas.annotations.*
 import org.springframework.http.MediaType
 import javax.servlet.ServletContext
 import org.apache.commons.io.FilenameUtils
 import org.springframework.web.multipart.MultipartFile
 
-//@Hidden
 @Api(value = "/videoUpload/", tags = ["Video"])
 class VideoUploadController extends BaseController {
 
-    @Value('${ISSUES_SECRET}')
-    private String issuesPassword
-
     def restService
-    def clientService
 
     //def beforeInterceptor = [action: this.checkHost()]
 
-    static allowedMethods = [uploadVideo: 'POST', submitIssue: 'POST', issueWebhook: 'POST']
-
-    static final String ISSUES_URL = "https://api.github.com"
-    static final String ISSUES_URI = "/repos/cormorack/feedback/issues"
-
-    /**
-     * Forwards to issue form
-     * @return
-     */
-    @ApiOperation(hidden = true)
-    def issueHookForm() {
-        [context: getAppContext()]
-    }
-
-    /**
-     * Creates a Github issue
-     * @return JSON response
-     */
-    @ApiOperation(
-            value = "Creates a Github Issue",
-            nickname = "issueWebhook",
-            consumes = "application/json",
-            produces = "application/json",
-            httpMethod = "POST"
-    )
-    @ApiResponses([
-            @ApiResponse(
-                    code = 200,
-                    message = "The POST call was successful"),
-            @ApiResponse(
-                    code = 405,
-                    message = "Method Not Allowed. Only POST is allowed"),
-            @ApiResponse(
-                    code = 404,
-                    message = "Method Not Found")
-    ])
-    @ApiImplicitParams([
-            @ApiImplicitParam(
-                    name = "name",
-                    paramType = "form",
-                    required = true,
-                    value = "Name",
-                    dataType = "string"),
-            @ApiImplicitParam(
-                    name = "body",
-                    paramType = "form",
-                    required = true,
-                    value = "Issue Description",
-                    dataType = "string"),
-            @ApiImplicitParam(
-                    name = "email",
-                    paramType = "form",
-                    required = true,
-                    value = "Email Address",
-                    dataType = "string"),
-            @ApiImplicitParam(
-                    name = "labels",
-                    paramType = "form",
-                    required = true,
-                    value = "Issue Label",
-                    dataType = "string")
-    ])
-    def issueWebhook() {
-
-        if (!params.body || !params.name || !params.email || !params.labels) {
-
-            Map data = ["message": "A required parameter is missing", "data": [] ]
-            Map results = ["succes": false, "data": data]
-            render results as JSON
-            return
-        }
-
-        String titleString = "${params.labels} feedback from ${params.name}"
-
-        Map paramMap = [title: titleString]
-
-        List labels = [params.labels]
-
-        paramMap.put("labels", labels)
-        paramMap.put("body", setDescription())
-
-        Map headerMap = ['Authorization': "token ${issuesPassword}", 'User-Agent': 'ooi-data-bot']
-
-        if (!clientService.postIssue(ISSUES_URL, ISSUES_URI, paramMap, headerMap)) {
-
-            log.error("An error occurred when submitting an issue")
-            Map data = ["message": "The operation could not be completed", "data": [] ]
-            Map results = ["succes": false, "data": data]
-            render results as JSON
-            return
-        }
-
-        Map data = ["message": "The form was sent successfully", "data": [] ]
-        Map results = ["succes": true, "data": data]
-
-        render results as JSON
-    }
-
-    /**
-     * Formats and fills the body with the name, email, label and body
-     * @param title
-     * @return formatted String
-     */
-    private String setDescription() {
-
-        String bodyString = """\
-        ## Overview
-        ${params.body}
-
-        ## Details
-        Sender: ${params.name}
-        Sender email: ${params.email}
-        Question type: ${params.labels}
-        """.stripIndent()
-    }
+    static allowedMethods = [uploadVideo: 'POST', submitIssue: 'POST']
 
     /**
      * Forwards to the video upload page
