+++ src/test/groovy/cavamedia/FeedbackControllerSpec.groovy
@@ -1,0 +1,18 @@
+package cavamedia
+
+import grails.testing.web.controllers.ControllerUnitTest
+import spock.lang.Specification
+
+class FeedbackControllerSpec extends Specification implements ControllerUnitTest<FeedbackController> {
+
+    def setup() {
+    }
+
+    def cleanup() {
+    }
+
+    void "test something"() {
+        expect:"fix me"
+            true == false
+    }
+}
